# swagger.yaml
openapi: 3.0.0
info:
  title: User API
  version: '1.0'
  description: An API for username/password creation and verification
paths:
  /api/user/register:
    post:
      tags:
        - user
      summary: Register User
      description: Creates a new user in the system
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username for the new user
                password:
                  type: string
                  description: The password for the new user
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /api/user/verify:
    post:
      tags:
        - user
      summary: Verify User
      description: Verifies a user's credentials (username and password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username for verification
                password:
                  type: string
                  description: The password for verification
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
components:
  responses:
    201:
      description: 'Success'
      content:
        text/plain: 
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates successful operation
              reason:
                type: string
                description: Optional reason for success or failure (if applicable)
    400:
      description: 'Bad Request'
      content:
        text/plain:
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates successful operation
              reason:
                type: string
                description: Optional reason for success or failure (if applicable)
    401:
      description: 'Unauthorized'
      content:
        text/plain: 
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates successful operation
              reason:
                type: string
                description: Optional reason for success or failure (if applicable)
    500:
      description: 'Internal server error'
      content:
        text/plain: 
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates successful operation
              reason:
                type: string
                description: Optional reason for success or failure (if applicable)